feat: export scrollToSection utility for library consumers
Extract scrollToSection from LoChaGroupList.vue into a standalone
utility at src/utils/scrollToSection.ts with an optional container
parameter, and export it from the library entry point.

Closes #102
